+++ CPP/src/BattleMonitor.cpp
@@ -25,16 +25,19 @@ static string t_inputStr;
 void TakeDmg(Player *player, int amt)
 {
     player->takeDmg(amt);
+    cout << player->getPlayerName() << "taking dmg: " << amt << endl;
 } //eo takeDmg
 
 void Heal(Player *player, int amt)
 {
     player->heal(amt);
+    cout << player->getPlayerName() << "healing: " << amt << endl;
 } //eo heal
 
 void IncreaseMaxHP(Player *player, int amt)
 {
     player->increaseMaxHP(amt);
+    cout << player->getPlayerName() << "incr HP: " << amt << endl;
 } //eo increaseMaxHP
 
 BattleMonitor::BattleMonitor() {}
@@ -125,6 +128,9 @@ void BattleMonitor::newBattle(Profile t_palyerProfile, Profile t_otherProfile, D
             displayBattleScreen(t_playerCardsPlayed, t_otherCardsPlayed);
             sleep(3);
 
+            this->m_CCRemainingPlayer = 3;
+            this->m_CCRemainingOther  = 3;
+            t_mustPlayCore = false;
             this->m_currPhaseNumber = 1;
             this->m_player.fillHand();
             this->m_otherPlayer.fillHand();
@@ -158,6 +164,7 @@ void BattleMonitor::newBattle(Profile t_palyerProfile, Profile t_otherProfile, D
 void BattleMonitor::procBattle(Card playerCard, Card otherCard)
 {
     queue<ActionPtr> actionQueue;
+    queue<ActionPtr> emptyQueue;
     queue<pair<Player*, int>> pairQueue;
 
     void (*ActionPtr) (Player*, int);
@@ -166,12 +173,25 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
     t_actionList[0] = pair<Action,Action>(playerCard.getFirstAction(), otherCard.getFirstAction());
     t_actionList[1] = pair<Action,Action>(playerCard.getSecondAction(), otherCard.getSecondAction());
 
-    for (uint8_t actionIdx = 0; actionIdx < TOTAL_CARD_ACTIONS; actionIdx++)
-    {
-        // PLAYER ACTION ONE: Attack - check if MS
-        // PLAYER ACTION ONE: Defend
-        // PLAYER ACTION ONE: Heal - check if LB
+    cout << "Player card: " << playerCard.getCardName() << endl;
+    cout << "Other  card: " << otherCard.getCardName()  << endl << endl;
 
+    cout << "Player Action 1: " << t_actionList[0].first << endl;
+    cout << "Other  Action 1: " << t_actionList[0].second << endl << endl;
+
+    cout << "Player Action 2: " << t_actionList[1].first << endl;
+    cout << "Other  Action 2: " << t_actionList[1].second << endl;
+
+    cout << "press any key to continue" << endl;
+    cin >> t_inputChar;
+
+    actionQueue = emptyQueue;
+
+    for (int actionIdx = 0; actionIdx < TOTAL_CARD_ACTIONS; actionIdx++)
+    {
+        cout << "processing action : " << actionIdx << endl;
+        cout << "t_actionList[actionIdx].first: " << t_actionList[actionIdx].first << endl;
+        cout << "t_actionList[actionIdx].second: " << t_actionList[actionIdx].second << endl;
         switch (t_actionList[actionIdx].first)
         {
         case ATTACK_ACTION:
@@ -205,22 +225,26 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                         pairQueue.push(pair<Player*, int>(&this->m_player,      1));
                         pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
                     }
+                    cout << "attack vs attack, push back 2" << endl;
                     ActionPtr = TakeDmg;
                     actionQueue.push(ActionPtr);
                     actionQueue.push(ActionPtr);
                     break;
                 case DEFEND_ACTION:
                     // Attack vs Block - Do nothing
+                    cout << "attack vs defend" << endl;
                     break;
                 case HEAL_ACTION:
                     // Mighty Swing vs Heal
                     if ((playerCard.getID().compare(to_string(ADV_MIGHTY_SWING_ID)) == 0))
                     {
+                        cout << "MS vs heal" << endl;
                         pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 3));
                     }
                     // Attack vs Heal
                     else
                     {
+                        cout << "attack vs heal" << endl;
                         pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
                     }
                     ActionPtr = TakeDmg;
@@ -228,6 +252,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     break;
                 case NO_ACTION:
                     // 2nd Action Attack vs 2nd Action None
+                    cout << "attack vs none" << endl;
                     pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
                     ActionPtr = TakeDmg;
                     actionQueue.push(ActionPtr);
@@ -249,7 +274,10 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     // Block vs Super Heal
                     if ((otherCard.getID().compare(to_string(ADV_SUPER_HEAL_ID)) == 0))
                     {
+                        cout << "block vs super heal" << endl;
                         pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 2));
+                        ActionPtr = Heal;
+                        actionQueue.push(ActionPtr);
                     }
                     // Block vs Life Boost
                     else if ((otherCard.getID().compare(to_string(ADV_LIFE_BOOST_ID)) == 0))
@@ -257,6 +285,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                         // First action
                         if (actionIdx == 0)
                         {
+                            cout << "block vs life boost action 1" << endl;
                             pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
                             ActionPtr = IncreaseMaxHP;
                             actionQueue.push(ActionPtr);
@@ -264,6 +293,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                         // Second action
                         else
                         {
+                            cout << "block vs life boost action 1=2" << endl;
                             pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
                             ActionPtr = Heal;
                             actionQueue.push(ActionPtr);
@@ -272,6 +302,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     // Block vs Heal
                     else
                     {
+                        cout << "block vs heal" << endl;
                         pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
                         ActionPtr = Heal;
                         actionQueue.push(ActionPtr);
@@ -298,6 +329,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     {
                         pairQueue.push(pair<Player*, int>(&this->m_player, 1));
                     }
+                    cout << "heal vs attack" << endl;
                     ActionPtr = TakeDmg;
                     actionQueue.push(ActionPtr);
                     break;
@@ -305,7 +337,10 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     // Super Heal vs Block
                     if ((playerCard.getID().compare(to_string(ADV_SUPER_HEAL_ID)) == 0))
                     {
+                        cout << "super heal vs block" << endl;
                         pairQueue.push(pair<Player*, int>(&this->m_player, 2));
+                        ActionPtr = Heal;
+                        actionQueue.push(ActionPtr);
                     }
                     // Life Boost vs Block
                     else if ((playerCard.getID().compare(to_string(ADV_LIFE_BOOST_ID)) == 0))
@@ -313,6 +348,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                         // First action
                         if (actionIdx == 0)
                         {
+                            cout << "life boost vs block 1st action" << endl;
                             pairQueue.push(pair<Player*, int>(&this->m_player, 1));
                             ActionPtr = IncreaseMaxHP;
                             actionQueue.push(ActionPtr);
@@ -320,6 +356,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                         // Second action
                         else
                         {
+                            cout << "life boost vs block 2nd action" << endl;
                             pairQueue.push(pair<Player*, int>(&this->m_player, 1));
                             ActionPtr = Heal;
                             actionQueue.push(ActionPtr);
@@ -328,6 +365,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     // Heal vs Block
                     else
                     {
+                        cout << "heal vs block" << endl;
                         pairQueue.push(pair<Player*, int>(&this->m_player, 1));
                         ActionPtr = Heal;
                         actionQueue.push(ActionPtr);
@@ -335,8 +373,9 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     break;
                 case HEAL_ACTION:
                     // Super Heal -  Player
+                    cout << "heal vs heal" << endl;
                     if ((playerCard.getID().compare(to_string(ADV_SUPER_HEAL_ID)) == 0))
-                    {
+                    { 
                         pairQueue.push(pair<Player*, int>(&this->m_player, 2));
                         ActionPtr = Heal;
                         actionQueue.push(ActionPtr);
@@ -401,12 +440,10 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     break;
                 case NO_ACTION:
                     // Heal vs 2nd Action
-                    if ((playerCard.getID().compare(to_string(ADV_LIFE_BOOST_ID)) == 0))
-                    {
-                        pairQueue.push(pair<Player*, int>(&this->m_player, 1));
-                        ActionPtr = Heal;
-                        actionQueue.push(ActionPtr);
-                    }
+                    cout << "heal vs no action" << endl;
+                    pairQueue.push(pair<Player*, int>(&this->m_player, 1));
+                    ActionPtr = Heal;
+                    actionQueue.push(ActionPtr);
                     break;
                 default:
                     // Do nothing
@@ -421,6 +458,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
              switch (t_actionList[actionIdx].second)
             {
                 case ATTACK_ACTION:
+                    cout << "no action vs attack" << endl;
                     // 2nd Action None vs Attack
                     pairQueue.push(pair<Player*, int>(&this->m_player, 1));
                     ActionPtr = TakeDmg;
@@ -431,14 +469,10 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
                     break;
                 case HEAL_ACTION:
                     // 2nd Action None vs Heal
-
-                    // 2nd Action heal should only be Life Boost
-                    if ((otherCard.getID().compare(to_string(ADV_LIFE_BOOST_ID)) == 0))
-                    {
-                        pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
-                        ActionPtr = TakeDmg;
-                        actionQueue.push(ActionPtr);
-                    }
+                    cout << "no action vs heal" << endl;
+                    pairQueue.push(pair<Player*, int>(&this->m_otherPlayer, 1));
+                    ActionPtr = Heal;
+                    actionQueue.push(ActionPtr);
                     break;
                 default:
                     // Do nothing
@@ -451,6 +485,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
 
     }//eof 
     
+    cout << "action queue size: " << actionQueue.size() << endl;
     // loop through action queue and exec actions
     while (actionQueue.empty() == false)
     {
@@ -458,7 +493,11 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
         pair<Player*, int> info = pairQueue.front();
         // Get action pointer and exec action
         ActionPtr = actionQueue.front();
+        cout << endl << "EXEC ACTION" << endl;
+        cout << "player: " << info.first->getPlayerName() << endl;
+
         ActionPtr(info.first, info.second);
+
         // Pop action off from queue
         pairQueue.pop();
         actionQueue.pop();
@@ -468,7 +507,7 @@ void BattleMonitor::procBattle(Card playerCard, Card otherCard)
 void BattleMonitor::displayBattleScreen(string t_playerCardsPlayed[5], string  t_otherCardsPlayed[5])
 {
     //clear console screen
-    system("cls");
+    //system("cls");
 
     cout << "Profile: " << this->m_player.getPlayerName() << endl;
     cout << "   Deck: " << this->m_player.getDeck()->getDeckName() << endl;
@@ -516,6 +555,11 @@ bool BattleMonitor::validatePlayerCardPick(char t_inputChar, bool *mustPlayCore)
     {
         *mustPlayCore = true;
     }
+    cout << "input validation returning false" << endl;
+    cout << "t_inputChar: " << t_inputChar << endl;
+    cout << "t_handIdx: " << t_handIdx << endl;
+    cout << "m_CCRemainingPlayer: " << m_CCRemainingPlayer << endl;
+    cout << "remainingPhases: " << remainingPhases << endl;
     return false;
 } //eo validateCardPick